Move the code than may use ACC into unifrac_cmp.cpp
