+++ sucpp/Makefile
@@ -90,14 +90,14 @@ endif
 
 CPPFLAGS += -Wall  -std=c++11 -pedantic -I. $(OPT) -fPIC -L$(CONDA_PREFIX)/lib
 
-test: tree.o test_su.cpp biom.o unifrac.o skbio_alt.o api.o
-	$(H5CXX) $(CPPFLAGS) $(EXEFLAGS) test_su.cpp -o test_su tree.o biom.o unifrac.o skbio_alt.o api.o -llz4 $(BLASLIB) -lpthread
-	$(H5CXX) $(CPPFLAGS) $(EXEFLAGS) test_ska.cpp -o test_ska skbio_alt.o tree.o biom.o unifrac.o api.o -llz4 $(BLASLIB) -lpthread
-	$(H5CXX) $(CPPFLAGS) $(EXEFLAGS) test_api.cpp -o test_api tree.o biom.o unifrac.o skbio_alt.o api.o -llz4 $(BLASLIB) -lpthread
+test: tree.o test_su.cpp biom.o unifrac_cmp.o unifrac.o skbio_alt.o api.o
+	$(H5CXX) $(CPPFLAGS) $(EXEFLAGS) test_su.cpp -o test_su tree.o biom.o unifrac_cmp.o unifrac.o skbio_alt.o api.o -llz4 $(BLASLIB) -lpthread
+	$(H5CXX) $(CPPFLAGS) $(EXEFLAGS) test_ska.cpp -o test_ska skbio_alt.o tree.o biom.o unifrac_cmp.o unifrac.o api.o -llz4 $(BLASLIB) -lpthread
+	$(H5CXX) $(CPPFLAGS) $(EXEFLAGS) test_api.cpp -o test_api tree.o biom.o unifrac_cmp.o unifrac.o skbio_alt.o api.o -llz4 $(BLASLIB) -lpthread
 
-main: tree.o biom.o unifrac.o cmd.o skbio_alt.o api.o
-	$(H5CXX) $(CPPFLAGS) $(EXEFLAGS) su.cpp -o ssu tree.o biom.o unifrac.o cmd.o skbio_alt.o api.o -lhdf5_cpp -llz4 $(BLASLIB) -lpthread
-	$(H5CXX) $(CPPFLAGS) $(EXEFLAGS) faithpd.cpp -o faithpd tree.o biom.o unifrac.o cmd.o skbio_alt.o api.o -lhdf5_cpp -llz4 $(BLASLIB) -lpthread
+main: tree.o biom.o unifrac_cmp.o unifrac.o cmd.o skbio_alt.o api.o
+	$(H5CXX) $(CPPFLAGS) $(EXEFLAGS) su.cpp -o ssu tree.o biom.o unifrac_cmp.o unifrac.o cmd.o skbio_alt.o api.o -lhdf5_cpp -llz4 $(BLASLIB) -lpthread
+	$(H5CXX) $(CPPFLAGS) $(EXEFLAGS) faithpd.cpp -o faithpd tree.o biom.o unifrac_cmp.o unifrac.o cmd.o skbio_alt.o api.o -lhdf5_cpp -llz4 $(BLASLIB) -lpthread
 	cp ssu ${PREFIX}/bin/
 	cp faithpd ${PREFIX}/bin/
 
@@ -115,8 +115,8 @@ rapi_test: main
 	echo LDFLAGS=-llz4 $(BLASLIB) >> ~/.R/Makevars
 	Rscript R_interface/rapi_test.R
 	
-api: tree.o biom.o unifrac.o cmd.o skbio_alt.o api.o
-	$(H5CXX) $(LDDFLAGS) -o libssu.so tree.o biom.o unifrac.o cmd.o skbio_alt.o api.o -lc -lhdf5_cpp -llz4 $(BLASLIB) -L$(PREFIX)/lib
+api: tree.o biom.o unifrac_cmp.o unifrac.o cmd.o skbio_alt.o api.o
+	$(H5CXX) $(LDDFLAGS) -o libssu.so tree.o biom.o unifrac_cmp.o unifrac.o cmd.o skbio_alt.o api.o -lc -lhdf5_cpp -llz4 $(BLASLIB) -L$(PREFIX)/lib
 	cp libssu.so ${PREFIX}/lib/
 
 capi_test: api
@@ -126,8 +126,12 @@ capi_test: api
 api.o: api.cpp api.hpp unifrac.hpp skbio_alt.hpp biom.hpp tree.hpp
 	$(H5CXX) $(CPPFLAGS) api.cpp -c -o api.o -fPIC
 
-unifrac.o: unifrac.cpp unifrac.hpp unifrac_task.cpp unifrac_task.hpp biom_interface.hpp tree.hpp
+unifrac.o: unifrac.cpp unifrac.hpp unifrac_internal.hpp unifrac_cmp.hpp biom_interface.hpp tree.hpp
 	$(CXX) $(CPPFLAGS) -c $< -o $@
+
+unifrac_cmp.o: unifrac_cmp.cpp unifrac_cmp.hpp unifrac_internal.hpp unifrac.hpp unifrac_task.cpp unifrac_task.hpp biom_interface.hpp tree.hpp
+	$(CXX) $(CPPFLAGS) -c $< -o $@
+
 
 %.o: %.cpp %.hpp
 	$(H5CXX) $(CPPFLAGS) -c $< -o $@
