+++ sucpp/unifrac.cpp
@@ -19,8 +19,8 @@
 #include <algorithm>
 #include <pthread.h>
 
-// embed in this file, to properly instantiate the templatized functions
-#include "unifrac_task.cpp"
+#include "unifrac_internal.hpp"
+#include "unifrac_cmp.hpp"
 
 
 static pthread_mutex_t printf_mutex;
@@ -68,6 +68,13 @@ void sig_handler(int signo) {
 }
 
 using namespace su;
+
+void su::try_report(const su::task_parameters* task_p, unsigned int k, unsigned int max_k) {
+  if(__builtin_expect(report_status[task_p->tid], false)) {
+    sync_printf("tid:%u\tstart:%u\tstop:%u\tk:%u\ttotal:%u\n", task_p->tid, task_p->start, task_p->stop, k, max_k);
+    report_status[task_p->tid] = false;
+  }
+}
 
 
 template<class TFloat>
@@ -137,38 +144,6 @@ TFloat* PropStack<TFloat>::pop(uint32_t node) {
 template class su::PropStack<float>;
 template class su::PropStack<double>;
 
-
-// Helper class with default constructor
-// The default is small enough to fit in L1 cache
-template<class TFloat>
-class PropStackFixed : public PropStack<TFloat> {
-  public:
-    static const uint32_t DEF_VEC_SIZE = 1024*sizeof(double)/sizeof(TFloat);
-
-    PropStackFixed() : PropStack<TFloat>(DEF_VEC_SIZE) {}
-};
-
-// Helper class that splits a large vec_size into several smaller chunks of def_size
-template<class TFloat> 
-class PropStackMulti {
-  protected:
-    const uint32_t vecsize;
-    std::vector<PropStackFixed<TFloat> > multi;
-
-  public:
-    PropStackMulti(uint32_t _vecsize) 
-    : vecsize(_vecsize)
-    , multi((vecsize + (PropStackFixed<TFloat>::DEF_VEC_SIZE-1))/PropStackFixed<TFloat>::DEF_VEC_SIZE) // round up
-    {}
-    ~PropStackMulti() {}
-
-    uint32_t get_num_stacks() const {return (vecsize + (PropStackFixed<TFloat>::DEF_VEC_SIZE-1))/PropStackFixed<TFloat>::DEF_VEC_SIZE;}
-
-    uint32_t get_start(uint32_t idx) const {return idx*PropStackFixed<TFloat>::DEF_VEC_SIZE;}
-    uint32_t get_end(uint32_t idx) const   {return std::min((idx+1)*PropStackFixed<TFloat>::DEF_VEC_SIZE, vecsize);}
-
-    PropStackFixed<TFloat> &get_prop_stack(uint32_t idx) {return multi[idx];}
-};
 
 double** su::deconvolute_stripes(std::vector<double*> &stripes, uint32_t n) {
     // would be better to just do striped_to_condensed_form
@@ -468,7 +443,7 @@ void progressbar(float progress) {
 }
 
 template<class TFloat>
-uint64_t initialize_sample_counts(TFloat*& _counts, const su::task_parameters* task_p, const biom_interface &table) {
+uint64_t initialize_sample_countsTT(TFloat*& _counts, const su::task_parameters* task_p, const biom_interface &table) {
     const unsigned int n_samples = task_p->n_samples;
     const uint64_t  n_samples_r = ((n_samples + UNIFRAC_BLOCK-1)/UNIFRAC_BLOCK)*UNIFRAC_BLOCK; // round up
     TFloat * counts = NULL;
@@ -492,10 +467,17 @@ uint64_t initialize_sample_counts(TFloat*& _counts, const su::task_parameters* t
    return n_samples_r;
 }
 
-void initialize_stripes(std::vector<double*> &dm_stripes,
-                        std::vector<double*> &dm_stripes_total,
-                        bool want_total,
-                        const su::task_parameters* task_p) {
+uint64_t su::initialize_sample_counts(double*& _counts, const su::task_parameters* task_p, const biom_interface &table) {
+   return initialize_sample_countsTT(_counts, task_p, table);
+}
+uint64_t su::initialize_sample_counts(float*& _counts, const su::task_parameters* task_p, const biom_interface &table) {
+   return initialize_sample_countsTT(_counts, task_p, table);
+}
+
+void su::initialize_stripes(std::vector<double*> &dm_stripes,
+                            std::vector<double*> &dm_stripes_total,
+                            bool want_total,
+                            const su::task_parameters* task_p) {
     int err = 0;
     for(unsigned int i = task_p->start; i < task_p->stop; i++){
         err = posix_memalign((void **)&dm_stripes[i], 4096, sizeof(double) * task_p->n_samples);
@@ -549,356 +531,26 @@ void su::faith_pd(biom_interface &table,
     }
 }
 
-template<class TaskT, class TFloat>
-void unifracTT(const biom_interface &table,
-               const BPTree &tree,
-               const bool want_total,
-               std::vector<double*> &dm_stripes,
-               std::vector<double*> &dm_stripes_total,
-               const su::task_parameters* task_p) {
-    int err;
-    // no processor affinity whenusing openacc or openmp
-
-    if(table.n_samples != task_p->n_samples) {
-        fprintf(stderr, "Task and table n_samples not equal\n");
-        exit(EXIT_FAILURE);
-    }
-    const unsigned int n_samples = task_p->n_samples;
-    const uint64_t  n_samples_r = ((n_samples + UNIFRAC_BLOCK-1)/UNIFRAC_BLOCK)*UNIFRAC_BLOCK; // round up
-
-
-    PropStackMulti<TFloat> propstack_multi(table.n_samples);
-
-    const unsigned int max_emb =  TaskT::RECOMMENDED_MAX_EMBS;
-
-    initialize_stripes(std::ref(dm_stripes), std::ref(dm_stripes_total), want_total, task_p);
-
-    TaskT taskObj(std::ref(dm_stripes), std::ref(dm_stripes_total),max_emb,task_p);
-
-    TFloat *lengths = NULL;
-    err = posix_memalign((void **)&lengths, 4096, sizeof(TFloat) * max_emb);
-    if(err != 0) {
-        fprintf(stderr, "posix_memalign(%d) failed: %d\n", sizeof(TFloat) * max_emb,  err);
-        exit(EXIT_FAILURE);
-    }
-#pragma acc enter data create(lengths[:max_emb])
-
-        /*
-         * The values in the example vectors correspond to index positions of an
-         * element in the resulting distance matrix. So, in the example below,
-         * the following can be interpreted:
-         *
-         * [0 1 2]
-         * [1 2 3]
-         *
-         * As comparing the sample for row 0 against the sample for col 1, the
-         * sample for row 1 against the sample for col 2, the sample for row 2
-         * against the sample for col 3.
-         *
-         * In other words, we're computing stripes of a distance matrix. In the
-         * following example, we're computing over 6 samples requiring 3
-         * stripes.
-         *
-         * A; stripe == 0
-         * [0 1 2 3 4 5]
-         * [1 2 3 4 5 0]
-         *
-         * B; stripe == 1
-         * [0 1 2 3 4 5]
-         * [2 3 4 5 0 1]
-         *
-         * C; stripe == 2
-         * [0 1 2 3 4 5]
-         * [3 4 5 0 1 2]
-         *
-         * The stripes end up computing the following positions in the distance
-         * matrix.
-         *
-         * x A B C x x
-         * x x A B C x
-         * x x x A B C
-         * C x x x A B
-         * B C x x x A
-         * A B C x x x
-         *
-         * However, we store those stripes as vectors, ie
-         * [ A A A A A A ]
-         *
-         * We end up performing N / 2 redundant calculations on the last stripe
-         * (see C) but that is small over large N.
-         */
-
-    unsigned int k = 0; // index in tree
-    const unsigned int max_k = (tree.nparens / 2) - 1;
-
-    const unsigned int num_prop_chunks = propstack_multi.get_num_stacks();
-    while (k<max_k) {
-          const unsigned int k_start = k;
-          unsigned int filled_emb = 0;
-
-          // chunk the progress to maximize cache reuse
-#pragma omp parallel for 
-          for (unsigned int ck=0; ck<num_prop_chunks; ck++) {
-            PropStack<TFloat> &propstack = propstack_multi.get_prop_stack(ck);
-            const unsigned int tstart = propstack_multi.get_start(ck);
-            const unsigned int tend = propstack_multi.get_end(ck);
-            unsigned int my_filled_emb = 0;
-            unsigned int my_k=k_start;
-
-            while ((my_filled_emb<max_emb) && (my_k<max_k)) {
-              const uint32_t node = tree.postorderselect(my_k);
-              my_k++;
-
-              TFloat *node_proportions = propstack.pop(node);
-              set_proportions_range(node_proportions, tree, node, table, tstart, tend, propstack);
-
-              if(task_p->bypass_tips && tree.isleaf(node))
-                  continue;
-
-              if (ck==0) { // they all do the same thing, so enough for the first to update the global state
-                lengths[filled_emb] = tree.lengths[node];
-                filled_emb++;
-              }
-              taskObj.embed_proportions_range(node_proportions, tstart, tend, my_filled_emb);
-              my_filled_emb++;
-            }
-            if (ck==0) { // they all do the same thing, so enough for the first to update the global state
-              k=my_k;
-            }
-          }
-
-          taskObj.sync_embedded_proportions(filled_emb);
-#ifdef _OPENACC
-          // lengths may be still in use in async mode, wait
-#pragma acc wait
-#pragma acc update device(lengths[:filled_emb])
-#endif
-          taskObj._run(filled_emb,lengths);
-          filled_emb=0;
-
-          if(__builtin_expect(report_status[task_p->tid], false)) {
-              sync_printf("tid:%d\tstart:%d\tstop:%d\tk:%d\ttotal:%d\n", task_p->tid, task_p->start, task_p->stop, k, max_k);
-              report_status[task_p->tid] = false;
-          }
-    }
-
-#pragma acc wait
-
-    if(want_total) {
-        const unsigned int start_idx = task_p->start;
-        const unsigned int stop_idx = task_p->stop;
-
-        TFloat * const dm_stripes_buf = taskObj.dm_stripes.buf;
-        const TFloat * const dm_stripes_total_buf = taskObj.dm_stripes_total.buf;
-
-#pragma acc parallel loop collapse(2) present(dm_stripes_buf,dm_stripes_total_buf)
-        for(unsigned int i = start_idx; i < stop_idx; i++)
-            for(unsigned int j = 0; j < n_samples; j++) {
-                unsigned int idx = (i-start_idx)*n_samples_r+j;
-                dm_stripes_buf[idx]=dm_stripes_buf[idx]/dm_stripes_total_buf[idx];
-                // taskObj.dm_stripes[i][j] = taskObj.dm_stripes[i][j] / taskObj.dm_stripes_total[i][j];
-            }
-        
-    }
-
-#pragma acc exit data delete(lengths[:max_emb])
-    free(lengths);
-}
-
+/*
+ * Defined in unifrac_cmd.cpp
 void su::unifrac(biom_interface &table,
                  BPTree &tree,
                  Method unifrac_method,
                  std::vector<double*> &dm_stripes,
                  std::vector<double*> &dm_stripes_total,
-                 const su::task_parameters* task_p) {
-    switch(unifrac_method) {
-        case unweighted:
-            unifracTT<SUCMP_NM::UnifracUnweightedTask<double>,double>(           table, tree, true,  dm_stripes,dm_stripes_total,task_p);
-            break;
-        case weighted_normalized:
-            unifracTT<SUCMP_NM::UnifracNormalizedWeightedTask<double>,double>(   table, tree, true,  dm_stripes,dm_stripes_total,task_p);
-            break;
-        case weighted_unnormalized:
-            unifracTT<SUCMP_NM::UnifracUnnormalizedWeightedTask<double>,double>( table, tree, false, dm_stripes,dm_stripes_total,task_p);
-            break;
-        case generalized:
-            unifracTT<SUCMP_NM::UnifracGeneralizedTask<double>,double>(          table, tree, true,  dm_stripes,dm_stripes_total,task_p);
-            break;
-        case unweighted_fp32:
-            unifracTT<SUCMP_NM::UnifracUnweightedTask<float >,float>(            table, tree, true,  dm_stripes,dm_stripes_total,task_p);
-            break;
-        case weighted_normalized_fp32:
-            unifracTT<SUCMP_NM::UnifracNormalizedWeightedTask<float >,float>(    table, tree, true,  dm_stripes,dm_stripes_total,task_p);
-            break;
-        case weighted_unnormalized_fp32:
-            unifracTT<SUCMP_NM::UnifracUnnormalizedWeightedTask<float >,float>(  table, tree, false, dm_stripes,dm_stripes_total,task_p);
-            break;
-        case generalized_fp32:
-            unifracTT<SUCMP_NM::UnifracGeneralizedTask<float >,float>(           table, tree, true,  dm_stripes,dm_stripes_total,task_p);
-            break;
-        default:
-            fprintf(stderr, "Unknown unifrac task\n");
-            exit(1);
-            break;
-    }
-}
+                 const su::task_parameters* task_p)
 
+*/
 
-template<class TaskT, class TFloat>
-void unifrac_vawTT(const biom_interface &table,
-                   const BPTree &tree,
-                          const bool want_total,
-                          std::vector<double*> &dm_stripes,
-                          std::vector<double*> &dm_stripes_total,
-                          const su::task_parameters* task_p) {
-    int err;
-    // no processor affinity whenusing openacc or openmp
-
-    if(table.n_samples != task_p->n_samples) {
-        fprintf(stderr, "Task and table n_samples not equal\n");
-        exit(EXIT_FAILURE);
-    }
-    const unsigned int n_samples = task_p->n_samples;
-    const uint64_t  n_samples_r = ((n_samples + UNIFRAC_BLOCK-1)/UNIFRAC_BLOCK)*UNIFRAC_BLOCK; // round up
-
-    PropStackMulti<TFloat> propstack_multi(table.n_samples);
-    PropStackMulti<TFloat> countstack_multi(table.n_samples);
-
-    const unsigned int max_emb = TaskT::RECOMMENDED_MAX_EMBS;
-
-    TFloat *sample_total_counts;
-
-    initialize_sample_counts<TFloat>(sample_total_counts, task_p, table);
-#pragma acc enter data copyin(sample_total_counts[:n_samples_r])
-    initialize_stripes(std::ref(dm_stripes), std::ref(dm_stripes_total), want_total, task_p);
-
-    TaskT taskObj(std::ref(dm_stripes), std::ref(dm_stripes_total), sample_total_counts, max_emb, task_p);
-
-    TFloat *lengths = NULL;
-    err = posix_memalign((void **)&lengths, 4096, sizeof(TFloat) * max_emb);
-    if(err != 0) {
-        fprintf(stderr, "posix_memalign(%d) failed: %d\n", sizeof(TFloat) * max_emb, err);
-        exit(EXIT_FAILURE);
-    }
-#pragma acc enter data create(lengths[:max_emb])
-
-    unsigned int k = 0; // index in tree
-    const unsigned int max_k = (tree.nparens / 2) - 1;
-
-    const unsigned int num_prop_chunks = propstack_multi.get_num_stacks();
-    while (k<max_k) {
-          const unsigned int k_start = k;
-          unsigned int filled_emb = 0;
-
-          // chunk the progress to maximize cache reuse
-#pragma omp parallel for 
-          for (unsigned int ck=0; ck<num_prop_chunks; ck++) {
-            PropStack<TFloat> &propstack = propstack_multi.get_prop_stack(ck);
-            PropStack<TFloat> &countstack = countstack_multi.get_prop_stack(ck);
-            const unsigned int tstart = propstack_multi.get_start(ck);
-            const unsigned int tend = propstack_multi.get_end(ck);
-            unsigned int my_filled_emb = 0;
-            unsigned int my_k=k_start;
-
-            while ((my_filled_emb<max_emb) && (my_k<max_k)) {
-              const uint32_t node = tree.postorderselect(my_k);
-              my_k++;
-
-              TFloat *node_proportions = propstack.pop(node);
-              TFloat *node_counts = countstack.pop(node);
-
-              set_proportions_range(node_proportions, tree, node, table, tstart, tend, propstack);
-              set_proportions_range(node_counts, tree, node, table, tstart, tend, countstack, false);
-
-              if(task_p->bypass_tips && tree.isleaf(node))
-                  continue;
-
-              if (ck==0) { // they all do the same thing, so enough for the first to update the global state
-                lengths[filled_emb] = tree.lengths[node];
-                filled_emb++;
-              }
-              taskObj.embed_range(node_proportions, node_counts, tstart, tend, my_filled_emb);
-              my_filled_emb++;
-            }
-            if (ck==0) { // they all do the same thing, so enough for the first to update the global state
-              k=my_k;
-            }
-          }
-
-#pragma acc wait
-#pragma acc update device(lengths[:filled_emb])
-          taskObj.sync_embedded(filled_emb);
-          taskObj._run(filled_emb,lengths);
-          filled_emb = 0;
-
-          if(__builtin_expect(report_status[task_p->tid], false)) {
-              sync_printf("tid:%d\tstart:%d\tstop:%d\tk:%d\ttotal:%d\n", task_p->tid, task_p->start, task_p->stop, k, max_k);
-              report_status[task_p->tid] = false;
-          }
-    }
-
-#pragma acc wait
-    if(want_total) {
-        const unsigned int start_idx = task_p->start;
-        const unsigned int stop_idx = task_p->stop;
-
-        TFloat * const dm_stripes_buf = taskObj.dm_stripes.buf;
-        const TFloat * const dm_stripes_total_buf = taskObj.dm_stripes_total.buf;
-
-#pragma acc parallel loop collapse(2) present(dm_stripes_buf,dm_stripes_total_buf)
-        for(unsigned int i = start_idx; i < stop_idx; i++)
-            for(unsigned int j = 0; j < n_samples; j++) {
-                unsigned int idx = (i-start_idx)*n_samples_r+j;
-                dm_stripes_buf[idx]=dm_stripes_buf[idx]/dm_stripes_total_buf[idx];
-                // taskObj.dm_stripes[i][j] = taskObj.dm_stripes[i][j] / taskObj.dm_stripes_total[i][j];
-            }
-
-    }
-
-
-#pragma acc exit data delete(lengths[:max_emb])
-#pragma acc exit data delete(sample_total_counts[:n_samples_r])
-    free(lengths);
-    free(sample_total_counts);
-}
-
+/*
+ *  Defined in unifrac_cmd.cpp
 void su::unifrac_vaw(biom_interface &table,
                      BPTree &tree,
                      Method unifrac_method,
                      std::vector<double*> &dm_stripes,
                      std::vector<double*> &dm_stripes_total,
-                     const su::task_parameters* task_p) {
-    switch(unifrac_method) {
-        case unweighted:
-            unifrac_vawTT<SUCMP_NM::UnifracVawUnweightedTask<double>,double>(           table, tree, true,  dm_stripes,dm_stripes_total,task_p);
-            break;
-        case weighted_normalized:
-            unifrac_vawTT<SUCMP_NM::UnifracVawNormalizedWeightedTask<double>,double>(   table, tree, true,  dm_stripes,dm_stripes_total,task_p);
-            break;
-        case weighted_unnormalized:
-            unifrac_vawTT<SUCMP_NM::UnifracVawUnnormalizedWeightedTask<double>,double>( table, tree, false, dm_stripes,dm_stripes_total,task_p);
-            break;
-        case generalized:
-            unifrac_vawTT<SUCMP_NM::UnifracVawGeneralizedTask<double>,double>(          table, tree, true,  dm_stripes,dm_stripes_total,task_p);
-            break;
-        case unweighted_fp32:
-            unifrac_vawTT<SUCMP_NM::UnifracVawUnweightedTask<float >,float >(           table, tree, true,  dm_stripes,dm_stripes_total,task_p);
-            break;
-        case weighted_normalized_fp32:
-            unifrac_vawTT<SUCMP_NM::UnifracVawNormalizedWeightedTask<float >,float >(   table, tree, true,  dm_stripes,dm_stripes_total,task_p);
-            break;
-        case weighted_unnormalized_fp32:
-            unifrac_vawTT<SUCMP_NM::UnifracVawUnnormalizedWeightedTask<float >,float >( table, tree, false, dm_stripes,dm_stripes_total,task_p);
-            break;
-        case generalized_fp32:
-            unifrac_vawTT<SUCMP_NM::UnifracVawGeneralizedTask<float >,float >(          table, tree, true,  dm_stripes,dm_stripes_total,task_p);
-            break;
-        default:
-            fprintf(stderr, "Unknown unifrac task\n");
-            exit(1);
-            break;
-    }
-}
+                     const su::task_parameters* task_p)
+ */
 
 template<class TFloat>
 void su::set_proportions(TFloat* __restrict__ props,
